Generalized Attribute Linking.
Generalized the attribute linking to a VAO, so that a vertex can have any number of attributes associated with it, using a function call to define all of them for each VAO. (for this example, colour was added). Also used a uniform to send information to the shader, aside from the buffers. (//note: added a simple scaling animation using uniforms just to keep it interesting)